cast some signed vs unsigned comparisons

+++ siod/editline.c
@@ -686,7 +686,7 @@ STATIC void clear_line()
     TTYputs(bol);
     for (i=screen_pos()/TTYwidth; i > 0; i--)
 	if (upline) TTYputs(upline);
-    for (i=0; i < strlen(Prompt); i++)
+    for (i=0; i < (int) strlen(Prompt); i++)
 	TTYput(' ');
     Point = 0;
     ceol();
@@ -1142,7 +1142,7 @@ STATIC STATUS meta()
     unsigned int	c;
     KEYMAP		*kp;
 
-    if ((c = TTYget()) == EOF)
+    if ((c = TTYget()) == (unsigned int) EOF)
 	return CSeof;
 #if	defined(ANSI_ARROWS)
     /* Also include VT-100 arrows. */
@@ -1158,7 +1158,7 @@ STATIC STATUS meta()
 #endif	/* defined(ANSI_ARROWS) */
 
     if (isdigit(c)) {
-	for (Repeat = c - '0'; (c = TTYget()) != EOF && isdigit(c); )
+	for (Repeat = c - '0'; (c = TTYget()) != (unsigned int) EOF && isdigit(c); )
 	    Repeat = Repeat * 10 + c - '0';
 	el_Pushed = 1;
 	el_PushBack = c;
@@ -1207,9 +1207,9 @@ STATIC STATUS TTYspecial(unsigned int c)
     if (ISMETA(c))
 	return CSdispatch;
 
-    if (c == rl_erase || c == DEL)
+    if (c == (unsigned int) rl_erase || c == DEL)
 	return bk_del_char();
-    if (c == rl_kill) {
+    if (c == (unsigned int) rl_kill) {
 	if (Point != 0) {
 	    for (i=screen_pos()/TTYwidth; i > 0; i--)
 		if (upline) TTYputs(upline);
@@ -1219,10 +1219,10 @@ STATIC STATUS TTYspecial(unsigned int c)
 	Repeat = NO_ARG;
 	return kill_line();
     }
-    if (c == rl_intr || c == rl_quit) {
+    if (c == (unsigned int) rl_intr || c == (unsigned int) rl_quit) {
 	Point = End = 0;
 	Line[0] = '\0';
-	if (c == rl_intr) 
+	if (c == (unsigned int) rl_intr) 
 	{
 	    el_intr_pending = 1;
 	    return CSdone;
@@ -1230,7 +1230,7 @@ STATIC STATUS TTYspecial(unsigned int c)
 	else 
 	    return redisplay();
     }
-    if (c == rl_eof && Point == 0 && End == 0)
+    if (c == (unsigned int) rl_eof && Point == 0 && End == 0)
 	return CSeof;
 
     return CSdispatch;
@@ -1244,7 +1244,7 @@ STATIC ECHAR *editinput()
     OldPoint = Point = Mark = End = 0;
     Line[0] = '\0';
 
-    while ((c = TTYget()) != EOF)
+    while ((c = TTYget()) != (unsigned int) EOF)
       {
 	switch (TTYspecial(c)) {
 	case CSdone:
@@ -1713,7 +1713,7 @@ STATIC STATUS quote()
 {
     unsigned int	c;
 
-    return (c = TTYget()) == EOF ? CSeof : insert_char((int)c);
+    return (c = TTYget()) == (unsigned int) EOF ? CSeof : insert_char((int)c);
 }
 
 STATIC STATUS wipe()
@@ -1744,9 +1744,9 @@ STATIC STATUS exchange()
     unsigned int	c;
 
     if ((c = TTYget()) != CTL('X'))
-	return c == EOF ? CSeof : ring_bell();
+	return c == (unsigned int) EOF ? CSeof : ring_bell();
 
-    if ((c = Mark) <= End) {
+    if ((c = Mark) <= (unsigned int) End) {
 	Mark = Point;
 	Point = c;
 	return CSmove;
@@ -1780,7 +1780,7 @@ STATIC STATUS move_to_char()
     int			i;
     ECHAR		*p;
 
-    if ((c = TTYget()) == EOF)
+    if ((c = TTYget()) == (unsigned int) EOF)
 	return CSeof;
     for (i = Point + 1, p = &Line[i]; i < End; i++, p++)
 	if (*p == c) {
